Extract extension loading into a helper module
Move the extension path resolution and fs.uploadZip into
stagehand-extension.ts so index.ts stays focused on the Stagehand flow.
