+++ pkg/templates/typescript/stagehand/index.ts
@@ -1,10 +1,8 @@
 import { Stagehand, localBrowser, type ModelName } from "@browserbasehq/stagehand";
 import { Kernel, type KernelContext } from "@onkernel/sdk";
-import { createReadStream } from "node:fs";
-import { dirname, join } from "node:path";
-import { fileURLToPath } from "node:url";
 // Stagehand v4 uses zod v4 schema types.
 import { z } from "zod";
+import { loadStagehandExtension } from "./stagehand-extension";
 
 const kernel = new Kernel();
 
@@ -27,16 +25,6 @@ const MODEL_API_KEY = process.env.MODEL_API_KEY ?? process.env.OPENAI_API_KEY;
 if (!MODEL_API_KEY) {
   throw new Error("MODEL_API_KEY (or OPENAI_API_KEY) is not set");
 }
-
-// Stagehand v4 runs as a Chrome extension alongside the browser. When
-// `localBrowser.connect` is called without an `extensionId`, Stagehand loads the
-// extension into the running browser over CDP via `Extensions.loadUnpacked`,
-// reading it from an absolute path on the *browser's* filesystem. Both the
-// unpacked extension and its archive ship inside the installed
-// @browserbasehq/stagehand package.
-const stagehandDist = dirname(fileURLToPath(import.meta.resolve("@browserbasehq/stagehand")));
-const STAGEHAND_EXTENSION_ZIP = join(stagehandDist, "assets/stagehand-extension.zip");
-const STAGEHAND_EXTENSION_DIR = join(stagehandDist, "extension");
 
 app.action<CompanyInput, TeamSizeOutput>(
   "teamsize-task",
@@ -64,15 +52,9 @@ app.action<CompanyInput, TeamSizeOutput>(
     let stagehand: Awaited<ReturnType<typeof Stagehand.create>> | undefined;
     let browser: Awaited<ReturnType<typeof localBrowser.connect>> | undefined;
     try {
-      // Mirror the Stagehand extension onto the running browser's filesystem at
-      // the exact path Stagehand loads it from, so the `Extensions.loadUnpacked`
-      // call made by `localBrowser.connect` (below) finds it.
-      await kernel.browsers.fs.uploadZip(kernelBrowser.session_id, {
-        dest_path: STAGEHAND_EXTENSION_DIR,
-        zip_file: createReadStream(STAGEHAND_EXTENSION_ZIP),
-      });
-
-      // No `extensionId`: Stagehand loads the mirrored extension over CDP.
+      // Load the Stagehand extension into the running browser. With no
+      // `extensionId`, `localBrowser.connect` then loads it over CDP.
+      await loadStagehandExtension(kernel, kernelBrowser.session_id);
       browser = await localBrowser.connect({
         cdpUrl: kernelBrowser.cdp_ws_url,
       });
